+++ tests/MagicOnion.Integration.Tests/Usings.cs
@@ -1,0 +1,6 @@
+global using Xunit;
+global using FluentAssertions;
+global using Moq;
+
+global using Grpc.Core;
+global using MessagePack;
